Move ResponseCollection into folder - #46
