+++ lib/trav3/response.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'response_collection'
+require_relative 'response/response_collection'
 require_relative 'response/response'
 require_relative 'response/success'
 require_relative 'response/request_error'
