+++ lib/trav3/response/response_collection.rb
